\name{PaPiSpan}
\alias{PaPiSpan}
\docType{data}
\title{
PaPiSpan
}
\description{
A dataframe giving the first and the last year of 44 series. The row names give the name of series.
}
\usage{data("PaPiSpan")}
\format{
  A data frame with 44 observations on the following 2 variables.
  \describe{
    \item{\code{first}}{a numeric vector giving the first year}
    \item{\code{last}}{a numeric vector giving the last year}
  }
}
\examples{
data(PaPiSpan)
head(PaPiSpan)
}
\keyword{datasets}
